Update fixtures to serve as a pytest plugin
[noissue]
