+++ functest_requirements.txt
@@ -1,3 +1,4 @@
+-e .
 pulp-smash @ git+https://github.com/pulp/pulp-smash.git
 orionutils
 pytest
